requirements.txt
torch_geometric
torch
numpy
tqdm
pyyaml